Move test helper 'console' inside './client' to avoid importing external files
